Refactor: Decouple settings from UserData in SyncDataUseCase
This commit refactors the `SyncDataUseCase` to make the `insertApplicationInfoGridItems` function more explicit and less dependent on the `UserData` object.

Previously, the function took the entire `UserData` object and destructured `experimentalSettings` and `homeSettings` internally. This has been changed to pass `ExperimentalSettings` and `HomeSettings` directly as parameters.

This change improves testability and clarifies the function's dependencies, as it no longer needs access to the entire `UserData` model. The call site within the `SyncDataUseCase` has been updated to pass these settings objects individually.